requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
sqlmodel>=0.0.8
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.5
email-validator>=1.1.3
pydantic-settings>=2.0.0
python-dotenv>=0.19.0
Faker>=24.0.0 